Moved IntelligentScissors to photo module.
