+++ modules/photo/perf/perf_intelligent_scissors.cpp
