+++ modules/photo/src/intelligent_scissors.cpp
