+++ modules/photo/test/test_intelligent_scissors.cpp
