+++ samples/cpp/CMakeLists.txt
@@ -24,8 +24,8 @@ if(NOT BUILD_EXAMPLES OR NOT OCV_DEPENDENCIES_FOUND)
   return()
 endif()
 
-set(DEPS_example_snippet_imgproc_segmentation opencv_core opencv_imgproc)
-set(DEPS_example_cpp_intelligent_scissors opencv_core opencv_imgproc opencv_imgcodecs opencv_highgui)
+set(DEPS_example_snippet_imgproc_segmentation opencv_core opencv_photo)
+set(DEPS_example_cpp_intelligent_scissors opencv_core opencv_photo opencv_imgcodecs opencv_highgui)
 
 project(cpp_samples)
 ocv_include_modules_recurse(${OPENCV_CPP_SAMPLES_REQUIRED_DEPS})
